refactor(dev): rename remix compiler module
