+++ packages/remix-dev/compiler/index.ts
@@ -3,4 +3,4 @@ export { type WatchOptions, watch } from "./watch";
 
 export { type CompileOptions, parseMode } from "./options";
 export { logCompileFailure } from "./onCompileFailure";
-export { CompileResult } from "./remixCompiler";
+export { CompileResult } from "./compiler";
